Fix bug with block deleting on Peer::finished()
When a Peer object received a finished() signal, it directly deleted all blocks it was waiting for.
Now it only deletes the blocks that ONLY it was waiting for (nobody else had this block in their m_blocksQueue)

+++ src/peer.cpp
@@ -353,7 +353,21 @@ void Peer::finished() {
 	m_handshakeTimeoutTimer.stop();
 	m_replyTimeoutTimer.stop();
 	for(auto block : m_blocksQueue) {
-		block->piece()->deleteBlock(block);
+		// Check if other peers have requested this block
+		bool othersHaveBlock = false;
+		for(Peer* peer : m_torrent->peers()) {
+			if(peer == this) {
+				continue;
+			}
+			if(peer->blocksQueue().contains(block)) {
+				othersHaveBlock = true;
+				break;
+			}
+		}
+		// Delete block if nobody has requested it
+		if(!othersHaveBlock) {
+			block->piece()->deleteBlock(block);
+		}
 	}
 	m_status = Created;
 	m_blocksQueue.clear();
